% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysql.R
\name{mysql_select}
\alias{mysql_select}
\title{查询相关数据}
\usage{
mysql_select(
  token = "9B75043D-8462-4D36-9B29-6D2EF4F3C2F3",
  sql = "select * from rds_crm_sales_saleOrder_api"
)
}
\arguments{
\item{token}{口令}

\item{sql}{sql}
}
\value{
返回值
}
\description{
查询相关数据
}
\examples{
mysql_select()
}
